% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapses.R
\name{banc_add_synapses}
\alias{banc_add_synapses}
\alias{banc_add_synapses.neuron}
\alias{banc_add_synapses.neuronlist}
\alias{banc_add_synapses.default}
\title{Add synapses to neuron objects}
\usage{
banc_add_synapses(x, ...)

\method{banc_add_synapses}{neuron}(
  x,
  id = NULL,
  connectors = NULL,
  size.threshold = 5,
  remove.autapses = TRUE,
  ...
)

\method{banc_add_synapses}{neuronlist}(
  x,
  id = NULL,
  connectors = NULL,
  size.threshold = 5,
  remove.autapses = TRUE,
  ...
)

\method{banc_add_synapses}{default}(
  x,
  id = NULL,
  connectors = NULL,
  size.threshold = 5,
  remove.autapses = TRUE,
  ...
)
}
\arguments{
\item{x}{A neuron object, neuronlist, or other object to add synapses to}

\item{...}{Additional arguments passed to methods, \code{nat::nlapply}}

\item{id}{The root ID of the neuron. If NULL, it uses the ID from the neuron object}

\item{connectors}{A dataframe of synaptic connections. If NULL, it retrieves the data}

\item{size.threshold}{Minimum size threshold for synapses to include}

\item{remove.autapses}{Whether to remove autapses (self-connections)}
}
\value{
An object of the same type as \code{x}, with synapses added
}
\description{
This function family adds synaptic data to neuron objects or neuron lists.
It retrieves synaptic connections and attaches them to the neuron object(s).
}
