% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave-tables.R
\name{banc_annotate_backbone_proofread}
\alias{banc_annotate_backbone_proofread}
\title{Annotate positions as backbone proofread}
\usage{
banc_annotate_backbone_proofread(
  positions,
  user_id,
  units = c("raw", "nm"),
  proofread = TRUE,
  datastack_name = NULL
)
}
\arguments{
\item{positions}{3D coordinates in BANC space}

\item{user_id}{Integer user ID for the annotation}

\item{units}{Character, coordinate units - either "raw" or "nm"}

\item{proofread}{Logical, whether to mark as proofread (default TRUE)}

\item{datastack_name}{Optional datastack name}
}
\value{
Data frame of added annotations
}
\description{
Mark specific positions as backbone proofread in the CAVE annotation system.
}
\examples{
\dontrun{
# Add an annotation to a point in raw voxel space
banc_annotate_backbone_proofread(c(117105, 240526, 5122), user_id = 355, units = "raw")

# Add an annotation to a point in nm
banc_annotate_backbone_proofread(c(468420, 962104 ,230490), user_id = 355, units = "nm")

# deannotate a point, only from points added with given user_id. Use user_id = NULL to remove from full pool
banc_deannotate_backbone_proofread(c(468420, 962104, 230490), user_id = 355, units = "nm")
}
}
