% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave.R
\name{banc_cave_client}
\alias{banc_cave_client}
\title{Low level access to banc's CAVE annotation infrastructure}
\usage{
banc_cave_client()
}
\value{
A reticulate R object wrapping the python CAVEclient.
}
\description{
Low level access to banc's CAVE annotation infrastructure
}
\examples{
\dontrun{
fcc=banc_cave_client()
tables=fcc$annotation$get_tables()
fcc$materialize$get_table_metadata(tables[1])
}
}
