% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave.R
\name{banc_cave_query}
\alias{banc_cave_query}
\title{Query banc tables in the CAVE annotation system}
\usage{
banc_cave_query(table, live = TRUE, ...)
}
\arguments{
\item{table}{The name of the table (or view, see views section) to query}

\item{live}{Whether to use live query mode, which updates any root ids to
their current value.}

\item{...}{Additional arguments passed to \code{\link{flywire_cave_query}}
\code{\link[fafbseg]{flywire_cave_query}}}
}
\value{
A data.frame
}
\description{
Query banc tables in the CAVE annotation system
}
\examples{
\donttest{
library(dplyr)
cell_info=banc_cave_query('cell_info')
cell_info \%>\%
  filter(tag2=='anterior-posterior projection pattern') \%>\%
  count(tag)
}
}
\seealso{
\code{\link[fafbseg]{flywire_cave_query}}
}
\concept{banc-cave}
