% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{banc_decapitate}
\alias{banc_decapitate}
\alias{banc_decapitate.NULL}
\alias{banc_decapitate.neuron}
\alias{banc_decapitate.neuronlist}
\alias{banc_decapitate.matrix}
\alias{banc_decapitate.data.frame}
\alias{banc_decapitate.mesh3d}
\alias{banc_decapitate.hxsurf}
\title{Subset points to be in the brain or in the VNC}
\usage{
banc_decapitate(x, y.cut = 325000, invert = FALSE, ...)

\method{banc_decapitate}{`NULL`}(x, y.cut = 325000, invert = FALSE, ...)

\method{banc_decapitate}{neuron}(x, y.cut = 325000, invert = FALSE, ...)

\method{banc_decapitate}{neuronlist}(x, y.cut = 325000, invert = FALSE, ...)

\method{banc_decapitate}{matrix}(x, y.cut = 325000, invert = FALSE, ...)

\method{banc_decapitate}{data.frame}(x, y.cut = 325000, invert = FALSE, ...)

\method{banc_decapitate}{mesh3d}(x, y.cut = 325000, invert = FALSE, ...)

\method{banc_decapitate}{hxsurf}(x, y.cut = 325000, invert = FALSE, ...)
}
\arguments{
\item{x}{an object with 3d points to be subsetted, e.g. an xyz matrix, a \code{neuron}, \code{neuronlist} or a \code{mesh3d} object.
Points must be in native BANC space, i.e. plottable inside \code{banc.surf}.}

\item{y.cut}{Numeric, the Y-axis cut point, in nanometers, in BANC space,
that separates the head from the neck and ventral nerve cord. For fitting to the
MANC data set, a cut height of \code{y.cut=5e05} seems good.}

\item{invert}{if \code{TRUE} returns brain points, if \code{FALSE} returns VNC points.}

\item{...}{Additional arguments passed to \code{\link{nlapply}} and then \code{\link{prune_vertices}}}
}
\value{
Remove points above or below the midsection of the neck connective of BANC.
}
\description{
Subset points to be in the brain or in the VNC
}
\examples{
\donttest{
# DNa02
m = banc_read_neuron_meshes("720575941478275714")
m.brain = banc_decapitate(m)
m.vnc = banc_decapitate(m, invert = TRUE)
}
\dontrun{
plot3d(m.brain, col = "red")
plot3d(m.vnc, col = "cyan")
plot3d(banc.surf, col = "grey", alpha = 0.1)
}
}
\seealso{
\code{\link{banc.surf}}
}
