% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{banc_ggneuron}
\alias{banc_ggneuron}
\title{Plot a neuron in the BANC connectomic dataset using ggplot2}
\usage{
banc_ggneuron(
  x,
  volume = NULL,
  info = NULL,
  view = c("main", "side", "front", "vnc", "vnc_side", "brain_side"),
  cols1 = c("turquoise", "navy"),
  cols2 = c("grey75", "grey50"),
  alpha = 0.5,
  title.col = "darkgrey",
  ...
)
}
\arguments{
\item{x}{A 'neuron', 'neuronlist', 'mesh3d', or 'hxsurf' object to be visualized.}

\item{volume}{A brain/neuropil volume to be plotted in grey, for context.
Defaults to NULL, no volume plotted.}

\item{info}{Optional. A string to be used as the plot title.}

\item{view}{A character string specifying the view orientation.
Options are "main", "side", "front", "vnc", "vnc_side", "brain_side".}

\item{cols1}{A vector of two colors for the lowest Z values. Default is c("turquoise", "navy").}

\item{cols2}{A vector of two colors for the highest Z values. Default is c("grey75", "grey50").}

\item{alpha}{Transparency of the neuron visualization. Default is 0.5.}

\item{title.col}{Color of the plot title. Default is "darkgrey".}

\item{...}{Additional arguments passed to geom_neuron().}
}
\value{
A ggplot object representing the visualized neuron.
}
\description{
This function visualizes a neuron or neuron-related object from the BANC connectomic dataset using ggplot2.
The only thing specific to the BANC data set is are the prreset 'view' angles.
}
\details{
This function is a wrapper around the ggneuron function, specifically tailored for the BANC dataset.
It applies a rotation matrix based on the specified view and uses predefined color schemes.
}
