% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{banc_ids}
\alias{banc_ids}
\title{Return a vector of banc root ids from diverse inputs}
\usage{
banc_ids(x, integer64 = NA)
}
\arguments{
\item{x}{A data.frame, URL or vector of ids}

\item{integer64}{Whether to return ids as \code{bit64::integer64} or
character vectors. Default value of NA leaves the ids unmodified.}
}
\value{
A vector of ids
}
\description{
Return a vector of banc root ids from diverse inputs
}
\examples{
banc_ids(data.frame(rootid="648518346474360770"))
}
\seealso{
Other banc-ids: 
\code{\link{banc_cellid_from_segid}()},
\code{\link{banc_islatest}()},
\code{\link{banc_latestid}()},
\code{\link{banc_leaves}()},
\code{\link{banc_rootid}()},
\code{\link{banc_xyz2id}()}
}
\concept{banc-ids}
