% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{banc_influence_arrow}
\alias{banc_influence_arrow}
\title{Read influence scores using Arrow dataset}
\usage{
banc_influence_arrow(parquet_dir, upstream_ids, downstream_ids)
}
\description{
Read influence scores using Arrow dataset
}
\keyword{internal}
