% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{banc_influence_duckdb}
\alias{banc_influence_duckdb}
\title{Read influence scores using DuckDB}
\usage{
banc_influence_duckdb(parquet_dir, upstream_ids, downstream_ids)
}
\description{
Read influence scores using DuckDB
}
\keyword{internal}
