% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{banc_islatest}
\alias{banc_islatest}
\title{Check if a banc root id is up to date}
\usage{
banc_islatest(x, timestamp = NULL, ...)
}
\arguments{
\item{x}{FlyWire rootids in any format understandable to
\code{\link[fafbseg]{ngl_segments}} including as \code{integer64}}

\item{timestamp}{(optional) argument to set an endpoint - edits after this
time will be ignored (see details).}

\item{...}{Additional arguments passed to \code{\link{flywire_islatest}}}
}
\description{
Check if a banc root id is up to date
}
\examples{
banc_islatest("720575941520182775")
}
\seealso{
Other banc-ids: 
\code{\link{banc_cellid_from_segid}()},
\code{\link{banc_ids}()},
\code{\link{banc_latestid}()},
\code{\link{banc_leaves}()},
\code{\link{banc_rootid}()},
\code{\link{banc_xyz2id}()}
}
\concept{banc-ids}
