% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{banc_latestid}
\alias{banc_latestid}
\alias{banc_updateids}
\title{Find the latest id for a banc root id}
\usage{
banc_latestid(
  rootid,
  sample = 1000L,
  cloudvolume.url = NULL,
  Verbose = FALSE,
  ...
)

banc_updateids(
  x,
  root.column = "root_id",
  supervoxel.column = "supervoxel_id",
  position.column = "position",
  use.cave = TRUE,
  serial = FALSE,
  ...
)
}
\arguments{
\item{rootid}{One ore more FlyWire rootids defining a segment (in any form
interpretable by \code{\link[fafbseg]{ngl_segments}})}

\item{sample}{An absolute or fractional number of supervoxel ids to map to
rootids or \code{FALSE} (see details).}

\item{cloudvolume.url}{URL for CloudVolume to fetch segmentation image data.
The default value of NULL chooses the flywire production segmentation
dataset.}

\item{Verbose}{When set to \code{TRUE} prints information about what fraction
of}

\item{...}{Additional arguments passed to \code{\link{flywire_latestid}}}

\item{x}{a \code{data.frame} with at least one of: \code{root_id}, \code{pt_root_id}, \code{supervoxel_id} and/or \code{pt_supervoxel_id}.
Supervoxels will be preferentially used to update the \code{root_id} column.
Else a vector of \code{BANC} root IDs.}

\item{root.column}{when \code{x} is a \code{data.frame}, the \code{root_id} column you wish to update}

\item{supervoxel.column}{when \code{x} is a \code{data.frame}, the \code{supervoxel_id} column you wish to use to update \code{root.column}}

\item{position.column}{when \code{x} is a \code{data.frame}, the \code{position} column with xyz values you wish to use to update \code{supervoxel.column}}

\item{use.cave}{read from the best established CAVE tables and join by \code{pt_supervoxel_id} to update \code{root_id}}

\item{serial}{if TRUE and x is a vector, calls \code{banc_updateids} on each ID in sequence to bufffer against connection failures. Slower.}
}
\description{
Find the latest id for a banc root id
}
\examples{
\dontrun{
banc_latestid("720575941520182775")
}
}
\seealso{
\code{\link{banc_islatest}}

Other banc-ids: 
\code{\link{banc_cellid_from_segid}()},
\code{\link{banc_ids}()},
\code{\link{banc_islatest}()},
\code{\link{banc_leaves}()},
\code{\link{banc_rootid}()},
\code{\link{banc_xyz2id}()}
}
\concept{banc-ids}
