% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners.R
\name{banc_partner_summary}
\alias{banc_partner_summary}
\title{Summarise the connectivity of banc neurons}
\usage{
banc_partner_summary(
  rootids,
  partners = c("outputs", "inputs"),
  threshold = 0,
  remove_autapses = TRUE,
  cleft.threshold = 0,
  datastack_name = NULL,
  ...
)
}
\arguments{
\item{rootids}{Character vector specifying one or more BANC rootids. As a
convenience this argument is passed to \code{\link{banc_ids}} allowing you
to pass in data.frames, BANC URLs or simple ids.}

\item{partners}{Whether to fetch input or output synapses or both.}

\item{threshold}{For \code{flywire_partner_summary} only return partners with
greater than this number of connections to the query neuron(s) (default of
0 returns all connections)}

\item{remove_autapses}{For \code{flywire_partner_summary} whether to remove
autapses (defaults to TRUE)}

\item{cleft.threshold}{A threshold for the cleft score calculated by Buhmann
et al 2019 (default 0, we have used 30-100 to increase specificity)}

\item{datastack_name}{An optional CAVE \code{datastack_name}. If unset a
sensible default is chosen.}

\item{...}{Additional arguments passed to \code{\link{pbsapply}}}
}
\value{
a data.frame
}
\description{
Summarise the connectivity of banc neurons
}
\details{
note that the rootids you pass in must be up to date. See example.
}
\examples{
\dontrun{
# NB id must be up to date
sample_id=banc_latestid("720575941478275714")
head(banc_partner_summary(sample_id))
head(banc_partner_summary(sample_id, partners='inputs'))
# get the latest id for an outdate
banc_partner_summary(banc_latestid("720575941478275714"))

## open banc/flywire scene containing top partners
library(dplyr)
banc_partner_summary(banc_latestid("720575941478275714"), partners='inputs') \%>\%
  slice_max(weight, n = 20) \%>\%
  banc_scene(open=TRUE)
}
}
\seealso{
\code{\link{flywire_partner_summary}}, \code{\link{banc_latestid}}
}
