% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{banc_read_neuroglancer_mesh}
\alias{banc_read_neuroglancer_mesh}
\title{Read BANC euroglancer meshes, e.g., ROI meshes}
\usage{
banc_read_neuroglancer_mesh(
  x = 1,
  url = paste0("https://www.googleapis.com/storage/v1/b/",
    "zetta_lee_fly_cns_001_kisuk/o/final\%2Fv2\%2F",
    "volume_meshes\%2Fmeshes\%2F{x}\%3A0.drc?alt=media",
    "&neuroglancer=610000b05b6497edcf20b78f29516970"),
  ...
)
}
\arguments{
\item{x}{the numeric identifier that specifies the mesh to read, defaults to \code{1} the BANC outline mesh.}

\item{url}{the URL that directs \code{bancr} to where BANC meshes are stored.}

\item{...}{arguments passed to \code{httr::GET}.}
}
\value{
a mesh3d object for the specified mesh.
}
\description{
Read BANC euroglancer meshes, e.g., ROI meshes
}
\examples{
\dontrun{
banc.mesh  <- banc_read_neuroglancer_mesh()
}
}
\seealso{
\code{\link{banc_read_neuron_meshes}}
}
