% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{banc_rootid}
\alias{banc_rootid}
\title{Find the root identifier of a banc neuron}
\usage{
banc_rootid(x, integer64 = FALSE, ...)
}
\arguments{
\item{x}{One or more FlyWire segment ids}

\item{integer64}{Whether to return ids as integer64 type (more compact but a
little fragile) rather than character (default \code{FALSE}).}

\item{...}{Additional arguments passed to \code{pbapply::pbsapply} and
eventually to Python \code{cv$CloudVolume} object.}
}
\value{
A vector of root ids (by default character)
}
\description{
Find the root identifier of a banc neuron
}
\examples{
\donttest{
banc_rootid("73186243730767724")
}
}
\seealso{
\code{\link{flywire_rootid}}

Other banc-ids: 
\code{\link{banc_cellid_from_segid}()},
\code{\link{banc_ids}()},
\code{\link{banc_islatest}()},
\code{\link{banc_latestid}()},
\code{\link{banc_leaves}()},
\code{\link{banc_xyz2id}()}
}
\concept{banc-ids}
