% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{banc_scene}
\alias{banc_scene}
\title{Return a sample Neuroglancer scene URL for BANC dataset}
\usage{
banc_scene(
  ids = NULL,
  open = FALSE,
  layer = NULL,
  url =
    "https://spelunker.cave-explorer.org/#!middleauth+https://global.daf-apis.com/nglstate/api/v1/6431332029693952"
)
}
\arguments{
\item{ids}{A set of root ids to include in the scene. Can also be a
data.frame.}

\item{open}{Whether to open the URL in your browser (see
\code{\link{browseURL}})}

\item{layer}{the segmentation layer for which \code{ids} intended. Defaults to 'segmentation proofreading',
but could point to another dataset layer.}

\item{url}{a spelunker neuroglancer URL.}
}
\value{
A character vector containing a single Neuroglancer URL (invisibly
when \code{open=TRUE}).
}
\description{
Return a sample Neuroglancer scene URL for BANC dataset
}
\details{
See
\href{https://banc-reconstruction.slack.com/archives/C01RZP5JH9C/p1616522511001900}{banc
slack} for details.
}
\examples{
\dontrun{
browseURL(banc_scene())
banc_scene(open=T)
banc_scene("720575941545083784", open=T)
}
}
\seealso{
\code{\link{bancsee}}
}
