% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudvolume.R
\name{banc_set_token}
\alias{banc_set_token}
\title{Set the token to be used to authenticate to banc autosegmentation resources}
\usage{
banc_set_token(token = NULL)
}
\arguments{
\item{token}{An optional token string. When missing you are prompted to
generate a new token via your browser.}
}
\value{
The path to the token file (invisibly)
}
\description{
Set the token to be used to authenticate to banc autosegmentation resources
}
