% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{banc_to_JRC2018F}
\alias{banc_to_JRC2018F}
\title{Transform Points between BANC Connectome and JRC2018F Template Brain}
\usage{
banc_to_JRC2018F(
  x,
  banc.units = c("nm", "um", "raw"),
  subset = NULL,
  inverse = FALSE,
  transform_file = NULL,
  method = c("tpsreg", "elastix", "navis_elastix_xform")
)
}
\arguments{
\item{x}{An object containing 3D points (must be compatible with nat::xyzmatrix).}

\item{banc.units}{Character string specifying the units of the BANC space data (input or output, depending on the inverse argument).
Must be one of "nm" (nanometers), "um", or "raw" (BANC raw banc.units). Default is "nm".}

\item{subset}{Optional. A logical vector or expression to subset the input object.}

\item{inverse}{Logical. If TRUE, performs the inverse transformation (JRC2018F to BANC).
Default is FALSE.}

\item{transform_file}{Optional. Path to a custom transform file. If NULL, uses default files.}

\item{method}{Character string specifying the transformation method.
Must be either "elastix" or "tpsreg". Default is "elastix".}
}
\value{
The input object with transformed 3D points.
}
\description{
This function transforms 3D points between the BANC (Buhmann et al. Adult Neural Connectome)
coordinate system and the D. melanogaster template brain JRC2018F coordinate system.
}
\details{
This function applies either an Elastix transform or a thin-plate spline registration to convert
points between the BANC and JRC2018F coordinate systems. It handles unit conversions as necessary.

The default transformation files are included with the package and are located in the
'inst/extdata/brain_240707' directory.
}
\examples{
\dontrun{
# Transform points from BANC to JRC2018F
transformed_points <- banc_to_JRC2018F(points, banc.units = "nm")

# Use a custom transform file
custom_transformed <- banc_to_JRC2018F(points, transform_file = "path/to/custom/transform.txt")

# Where the default transform files are located:
banc_to_JRC2018F_file <- system.file(file.path("extdata","brain_240707"),
"BANC_to_template.txt", package="bancr")
JRC2018F_to_banc_file <- system.file(file.path("extdata","brain_240707"),
"template_to_BANC.txt", package="bancr")
}

}
\seealso{
\code{\link{elastix_xform}} for the underlying Elastix transformation function.
\code{\link{banc_raw2nm}} and \code{\link{banc_nm2raw}} for unit conversion functions.
}
