% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroglancer-utils.R
\name{banc_upload_mesh}
\alias{banc_upload_mesh}
\title{Upload Mesh to Google Cloud Storage for BANC Neuroglancer}
\usage{
banc_upload_mesh(
  mesh,
  mesh_id,
  vol = "precomputed://gs://lee-lab_brain-and-nerve-cord-fly-connectome/volume_meshes/",
  compress = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{mesh}{Character or mesh3d object. Either a path to the .obj file or an R mesh object.}

\item{mesh_id}{Integer. ID for the mesh.}

\item{vol}{Character. URL for the precomputed volume.}

\item{compress}{Logical. Whether to compress the mesh. Default is \code{TRUE}.}

\item{overwrite}{Logical. Whether or not to overwrite an extant file of the same name and path at
the location specified by \code{vol}+\code{mesh_id}. Default is \code{FALSE}.}
}
\value{
Invisible \code{NULL}. The function is called for it upload effect.
}
\description{
This function uploads a mesh.obj file to a Google Cloud Storage bucket
that can be accessed by BANC neuroglancer. It uses reticulate to call
Python code and manages a conda environment.
}
\details{
The underlying Python code is based on the bikinibottom package:
\url{https://github.com/jasper-tms/bikini-bottom/tree/main}

For this function to work the global environment variable \code{GOOGLE_APPLICATION_CREDENTIALS} to
something like \code{"$HOME/.config/gcloud/legacy_credentials/firstname.secondname@gmail.com/adc.json"}.
This file can be obtains by installing \code{gsutil} and running \verb{gsutil init} in the terminal.
}
\examples{
\dontrun{
banc_upload_mesh(
mesh = "/Users/abates/projects/flyconnectome/bancpipeline/deformetrica/obj/banc_brain_neuropil.obj",
mesh_id = 3,
vol = "precomputed://gs://lee-lab_brain-and-nerve-cord-fly-connectome/volume_meshes/meshes"
)
}
}
\references{
bikinibottom package: \url{https://github.com/jasper-tms/bikini-bottom/tree/main}
}
