% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banc_users}
\alias{banc_users}
\title{User information (name + CAVE ID) for active BANC users}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 206 rows and 3 columns.
}
\usage{
banc_users
}
\description{
The purpose of this table is to map CAVE users IDs to names, in order to credit annotation work done in BANC CAVE.
This information is based on \href{https://docs.google.com/spreadsheets/d/1UFmeWr2uF9jTLVMw3bD6nM3ejM-b-HDZz6sQBPTEoZ8/edit?gid=1163959922#gid=1163959922}{google sheet}.
}
\examples{
\dontrun{
View(banc_users)
}
}
\keyword{datasets}
