% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{banc_view}
\alias{banc_view}
\alias{banc_side_view}
\alias{banc_front_view}
\alias{banc_vnc_view}
\alias{banc_vnc_side_view}
\alias{banc_brain_side_view}
\title{Set Default View for BANC EM Dataset}
\usage{
banc_view()

banc_side_view()

banc_front_view()

banc_vnc_view()

banc_vnc_side_view()

banc_brain_side_view()
}
\value{
This function is called for its side effect of changing the rgl viewpoint.
It does not return a value.
}
\description{
This function sets a default view for visualizing the 'BANC' Electron Microscopy (EM) dataset
using the rgl package. It adjusts the viewpoint to a specific orientation and zoom level
that is optimal for viewing this particular dataset.
}
\details{
The function uses \code{rgl::rgl.viewpoint()} to set a predefined user matrix and zoom level.
This matrix defines the rotation and translation of the view, while the zoom parameter
adjusts the scale of the visualization.
}
\note{
This function assumes that an rgl device is already open and that the BANC EM dataset
has been plotted. It will not create a new plot or open a new rgl device.
}
\examples{
\dontrun{
# Assuming you have already plotted your BANC EM data
banc_view()
}

}
\seealso{
\code{\link[rgl]{rgl.viewpoint}} for more details on setting viewpoints in rgl.
}
