% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{banc_voxdims}
\alias{banc_voxdims}
\alias{banc_nm2raw}
\alias{banc_raw2nm}
\title{Handle raw and nm calibrated banc coordinates}
\usage{
banc_voxdims(url = choose_banc(set = FALSE)[["fafbseg.sampleurl"]])

banc_nm2raw(x, vd = banc_voxdims())

banc_raw2nm(x, vd = banc_voxdims())
}
\arguments{
\item{url}{Optional neuroglancer URL containing voxel size. Defaults to
\code{getOption("fafbseg.sampleurl")} as set by
\code{\link{choose_banc}}.}

\item{x}{3D coordinates in any form compatible with \code{\link{xyzmatrix}}}

\item{vd}{The voxel dimensions in nm. Expert use only. Normally found
automatically.}
}
\value{
For \code{banc_voxdims} A 3-vector

for \code{banc_raw2nm} and \code{banc_nm2raw} an Nx3 matrix of
coordinates
}
\description{
\code{banc_voxdims} returns the image voxel dimensions which
are normally used to scale between \bold{raw} and \bold{nm} coordinates.
}
\details{
relies on nat >= 1.10.4
}
\examples{
banc_voxdims()
banc_raw2nm(c(159144, 22192, 3560))
banc_raw2nm('159144 22192 3560')
\dontrun{
banc_nm2raw(clipr::read_clip())
}
}
