% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bancr-package.R
\docType{package}
\name{bancr-package}
\alias{bancr}
\alias{bancr-package}
\title{bancr: R Client Access to the Brain And Nerve Cord (BANC) Dataset}
\description{
Access to the BANC 'righty' EM dataset, focussing on support for BANC autosegmentation. Organised as a wrapper around the fafbseg package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flyconnectome/bancr}
  \item \url{https://flyconnectome.github.io/bancr/}
  \item Report bugs at \url{https://github.com/flyconnectome/bancr/issues}
}

}
\author{
\strong{Maintainer}: Alexnder Bates \email{alexander.shakeel.bates@gmail.com} (\href{https://orcid.org/0000-0002-1195-0445}{ORCID})

Authors:
\itemize{
  \item Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})
}

}
\keyword{internal}
