% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banc-table.R
\name{banctable_query}
\alias{banctable_query}
\alias{banc_set_token}
\alias{banctable_login}
\alias{banctable_update_rows}
\title{Read and write to the seatable for draft BANC annotations}
\usage{
banctable_query(
  sql = "SELECT * FROM banc_meta",
  limit = 100000L,
  base = NULL,
  python = FALSE,
  convert = TRUE,
  ac = NULL
)

banc_set_token(token = NULL)

banctable_login(
  url = "https://cloud.seatable.io/",
  token = Sys.getenv("BANCTABLE_TOKEN", unset = NA_character_)
)

banctable_update_rows(
  df,
  table,
  base = NULL,
  append_allowed = TRUE,
  chunksize = 1000L,
  ...
)
}
\arguments{
\item{sql}{A SQL query string. See examples and
\href{https://seatable.github.io/seatable-scripts/python/query/}{seatable
docs}.}

\item{limit}{An optional limit, which only applies if you do not specify a
limit directly in the \code{sql} query. By default seatable limits SQL
queries to 100 rows. We increase the limit to 100000 rows by default.}

\item{base}{Character vector specifying the \code{base}}

\item{python}{Logical. Whether to return a Python pandas DataFrame. The default of FALSE returns an R data.frame}

\item{convert}{Expert use only: Whether or not to allow the Python seatable
module to process raw output from the database. This is is principally for
debugging purposes. NB this imposes a requirement of seatable_api >=2.4.0.}

\item{ac}{A seatable connection object as returned by \code{banctable_login}.}

\item{token}{normally retrieved from \code{BANCTABLE_TOKEN} environment
variable.}

\item{url}{Optional URL to the server}

\item{df}{A data.frame containing the data to upload including an \verb{_id}
column that can identify each row in the remote table.}

\item{table}{Character vector specifying a table foe which you want a
\code{base} object.}

\item{append_allowed}{Logical. Whether rows without row identifiers can be appended.}

\item{chunksize}{To split large requests into smaller ones with max this many rows.}

\item{...}{Additional arguments passed to pbsapply which might include cl=2 to specify a number of parallel jobs to run.}

\item{user, pwd}{banctable user and password used by \code{banctable_set_token}
to obtain a token}
}
\value{
a \code{data.frame} of results. There should be 0 rows if no rows
matched query.
}
\description{
These functions use the logic and wrap some code
from the \verb{flytable_.*} functions in the \code{fafbseg} R package.
\code{banctable_set_token} will obtain and store a permanent
seatable user-level API token.
\code{banctable_query} performs a SQL query against a banctable
database. You can omit the \code{base} argument unless you have tables of
the same name in different bases.
\code{banctable_base} returns a \code{base} object (equivalent to
a mysql database) which allows you to access one or more tables, logging in
to the service if necessary. The returned base object give you full access
to the Python
\href{https://seatable.github.io/seatable-scripts/python/base/}{\code{Base}}
API allowing a range of row/column manipulations.
\code{banctable_update_rows} updates existing rows in a table, returning TRUE on success.
}
\examples{
\dontrun{
banc_set_token(user="my_gmail.com",
banc.meta <- banctable_query()
}
}
\seealso{
\code{\link{fafbseg::flytable_query}}
}
