% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{check_if_possible}
\alias{check_if_possible}
\title{Check if Elastix Transform is Possible}
\usage{
check_if_possible(file, on_error = "raise")
}
\arguments{
\item{file}{Path to the Elastix transform file.}

\item{on_error}{Action to take on error: "raise" to stop execution, or any other value to return an error message.}
}
\value{
NULL if file exists, or an error message if the file doesn't exist and on_error is not "raise".
}
\description{
Verifies if the specified transform file exists.
}
\keyword{internal}
