% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{choose_banc}
\alias{choose_banc}
\alias{with_banc}
\title{Choose or (temporarily) use the banc autosegmentation}
\usage{
choose_banc(set = TRUE)

with_banc(expr)
}
\arguments{
\item{set}{Whether or not to permanently set the banc autosegmentation as the
default for \code{\link{fafbseg}} functions.}

\item{expr}{An expression to evaluate while banc is the default
autosegmentation}
}
\value{
If \code{set=TRUE} a list containing the previous values of the
relevant global options (in the style of \code{\link{options}}. If
\code{set=FALSE} a named list containing the option values.
}
\description{
Choose or (temporarily) use the banc autosegmentation
}
\details{
\code{bancr} inherits a significant amount of infrastructure from
the \code{\link{fafbseg}} package. This has the concept of the
\emph{active} autosegmentation, which in turn defines one or more R options
containing URIs pointing to voxel-wise segmentation, mesh etc data. These
are normally contained within a single neuroglancer URL which points to
multiple data layers. For banc this is the neuroglancer scene returned by
\code{\link{banc_scene}}.
}
\examples{
\dontrun{
choose_banc()
options()[grep("^fafbseg.*url", names(options()))]
}
\dontrun{
with_banc(fafbseg::flywire_islatest('648518346498254576'))
}
\dontrun{
with_banc(fafbseg::flywire_latestid('648518346498254576'))
with_banc(fafbseg::flywire_latestid('648518346494405175'))
}
}
