% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{elastix_xform}
\alias{elastix_xform}
\title{Perform Elastix Transform on 3D Points}
\usage{
elastix_xform(points, transform_file, copy_files = c(), return_logs = FALSE)
}
\arguments{
\item{points}{A matrix with 3 columns or a data frame with x, y, z columns representing 3D points.}

\item{transform_file}{Path to the Elastix transform file.}

\item{copy_files}{Vector of additional file paths to copy to the temporary directory.}

\item{return_logs}{Logical, if TRUE, returns the Elastix log instead of transformed points.}
}
\value{
A matrix of transformed 3D points, or Elastix logs if return_logs is TRUE.
}
\description{
This function applies an Elastix spatial transform to a set of 3D points.
}
\details{
This function requires Elastix to be installed and added to the system PATH.
It creates a temporary directory for processing, applies the Elastix transform,
and cleans up afterwards.
}
\examples{
\dontrun{
points <- matrix(rnorm(30), ncol = 3)
transformed_points <- elastix_xform(points, "path/to/transform.txt")
}

}
