% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{ggneuron}
\alias{ggneuron}
\title{Create a ggplot2 Visualisation of Neuron Objects}
\usage{
ggneuron(
  x,
  volume = NULL,
  info = NULL,
  rotation_matrix = NULL,
  low1 = "turquoise",
  high1 = "navy",
  low2 = "grey75",
  high2 = "grey50",
  alpha = 0.5,
  title.col = "darkgrey",
  ...
)
}
\arguments{
\item{x}{A 'neuron', 'neuronlist', 'mesh3d', or 'hxsurf' object to be visualized.}

\item{volume}{a brain/neuropil volume to be plotted in grey, for context.
Defaults to NULL, no volume plotted.}

\item{info}{Optional. A string to be used as the plot title.}

\item{rotation_matrix}{An optional 4x4 rotation matrix to apply to the neuron coordinates.}

\item{low1, low2}{Color for the lowest Z values. Default is "turquoise".}

\item{high1, high2}{Color for the highest Z values. Default is "navy".}

\item{alpha}{Transparency of the neuron visualization. Default is 0.5.}

\item{title.col}{Color of the plot title. Default is "darkgrey".}

\item{...}{Additional arguments passed to geom_neuron().}
}
\value{
A ggplot2 object representing the neuron visualization.
}
\description{
This function creates a complete ggplot2 visualization for neuron objects,
including 'neuron', 'neuronlist', 'mesh3d', and 'hxsurf' objects. It sets up
a minimal theme and applies consistent styling to the plot.
}
\details{
This function wraps around geom_neuron() to create a complete plot with a
consistent, minimal theme. It removes axes, legends, and other extraneous
elements to focus on the neuron visualization itself.
}
\examples{
\dontrun{
# Visualize the banc volume
ggneuron(banc_neuropil.surf, banc.surf)

# Visualize the banc brain neuropil
ggneuron(banc_neuropil.surf,
rotation_matrix = bancr:::banc_rotation_matrices[["front"]])

# See constituents of a neuronlist
ggneuron(nat::Cell07PNs, volume = nat::MBL.surf,
root = 10, high = "red", low = "black")
}

}
\seealso{
\code{\link{geom_neuron}} for the underlying geom used in this function.
}
