% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{navis_elastix_xform}
\alias{navis_elastix_xform}
\title{Apply Elastix Transform using Navis}
\usage{
navis_elastix_xform(x, transform_path)
}
\arguments{
\item{x}{Matrix or data frame of 3D points.}

\item{transform_path}{Path to the Elastix transform file, usually a \code{.txt} file.}
}
\value{
A matrix of transformed 3D points.
}
\description{
Applies an Elastix transform to 3D points using the Navis Python library.
}
\details{
This function requires the reticulate R package and the Navis Python library.
}
\examples{
\dontrun{
points <- matrix(rnorm(30), ncol = 3)
transformed_points <- navis_elastix_xform(points, "path/to/transform.txt")
}

}
