% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{read_elastix_output_file}
\alias{read_elastix_output_file}
\title{Read Elastix Output File}
\usage{
read_elastix_output_file(filepath)
}
\arguments{
\item{filepath}{Path to the Elastix output file.}
}
\value{
A matrix of transformed 3D points.
}
\description{
Reads and parses the output file produced by Elastix transform.
}
\keyword{internal}
