% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{write_elastix_input_file}
\alias{write_elastix_input_file}
\title{Write Elastix Input File}
\usage{
write_elastix_input_file(points, filepath)
}
\arguments{
\item{points}{Matrix of 3D points.}

\item{filepath}{Path where the input file should be written.}
}
\description{
Writes 3D points to a file in the format required by Elastix.
}
\keyword{internal}
