% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banc.surf}
\alias{banc.surf}
\title{Simplified tissue surface of BANC}
\format{
An object of class \code{hxsurf} (inherits from \code{list}) of length 4.
}
\usage{
banc.surf
}
\description{
This is unsymmetrical and not a normalised version of the mesh.
It is the outline of the dataset in native voxel space.
}
\examples{
\dontrun{
# Depends on nat
library(nat)
rgl::wire3d(banc.surf)
}
}
\keyword{datasets}
