% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banc.surf}
\alias{banc.surf}
\alias{banc_neuropil.surf}
\alias{banc_brain_neuropil.surf}
\alias{banc_vnc_neuropil.surf}
\alias{banc_neck_connective.surf}
\alias{banc_neuropils.surf}
\alias{banc_al.surf}
\title{Simplified tissue and neuropil surfaces for BANC}
\format{
An object of class \code{hxsurf} (inherits from \code{list}) of length 4.

An object of class \code{hxsurf} (inherits from \code{list}) of length 4.

An object of class \code{hxsurf} (inherits from \code{list}) of length 4.

An object of class \code{hxsurf} (inherits from \code{list}) of length 4.

An object of class \code{hxsurf} (inherits from \code{list}) of length 4.

An object of class \code{hxsurf} (inherits from \code{list}) of length 4.

An object of class \code{hxsurf} of length 4.
}
\usage{
banc.surf

banc_neuropil.surf

banc_brain_neuropil.surf

banc_vnc_neuropil.surf

banc_neck_connective.surf

banc_neuropils.surf

banc_al.surf
}
\description{
\code{banc.surf} is unsymmetrical and not a normalised version of the mesh.
It is the outline of the dataset in nanometers.\code{banc_neuropil.surf} represents the synaptic neuropil.
Built from the BANC synapse cloud, but not optimised to include 100\% of bona fide presynapses.
\code{banc_neuropils.surf} contains the standard Ito et al., 2014 brain neuropil volumes transformed into BANC space.
\code{banc_al.surf} contains the standard Bates and Schlegel et al, 2021 right antennal lobe glomeruli brain neuropil volumes transformed into BANC space.
}
\examples{
\dontrun{
# Depends on nat
library(nat)
rgl::wire3d(banc.surf)
}
}
\keyword{datasets}
