% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapses.R
\name{banc_all_synapses}
\alias{banc_all_synapses}
\title{Download all of the BANC synapses as a .sqlite file that you can read lazily from later}
\usage{
banc_all_synapses(
  path = "gs://zetta_lee_fly_cns_001_synapse/240623_run/assignment/final_edgelist.df",
  overwrite = FALSE,
  n_max = 2000
)
}
\arguments{
\item{path}{The google storage path to the desired synapses file. Read using \code{readr::read_csv}.}

\item{overwrite}{Logical, whether or not to overwrite an extant \code{banc_data.sqlite} file.}

\item{n_max}{Numeric, the maximum number of rows ot read from \code{path} if you just want to see
a taster of the file.}
}
\value{
a data.frame
}
\description{
Download all of the BANC synapses as a .sqlite file that you can read lazily from later
}
\details{
Downloads all automatic Zetta.ai synapse detections for the BANC and saves them as
a \code{banc_data.sqlite} file. Once this is done, in the future the function will read from this file
lazilty so as not to throw the whole thing into system memory.
}
\examples{
\dontrun{
syns <- banc_all_synapses()
}
}
\seealso{
\code{\link{banc_partner_summary}}, \code{\link{banc_partners}}
}
