% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave-tables.R
\name{banc_cave_tables}
\alias{banc_cave_tables}
\alias{banc_nuclei}
\alias{banc_cell_info}
\alias{banc_cell_ids}
\alias{banc_neck_connective_neurons}
\alias{banc_peripheral_nerves}
\alias{banc_backbone_proofread}
\title{Read BANC CAVE-tables, good sources of metadata}
\usage{
banc_cave_tables(
  datastack_name = getOption("fafbseg.cave.datastack_name"),
  select = NULL
)

banc_nuclei(
  rootids = NULL,
  nucleus_ids = NULL,
  table = c("somas_v1a", "somas_v1b"),
  rawcoords = FALSE,
  ...
)

banc_cell_info(rootids = NULL, rawcoords = FALSE)

banc_cell_ids(rootids = NULL)

banc_neck_connective_neurons(
  rootids = NULL,
  table = c("neck_connective_y92500", "neck_connective_y121000")
)

banc_peripheral_nerves(rootids = NULL)

banc_backbone_proofread(rootids = NULL)
}
\arguments{
\item{datastack_name}{Defaults to "brain_and_nerve_cord". See https://global.daf-apis.com/info/ for other options.}

\item{select}{A regex term for the name of the table you want}

\item{table}{Possible alternative tables for the sort of data frame the function returns. One must be chosen.}

\item{...}{Additional arguments passed to
\code{fafbseg::\link{flywire_cave_query}}}
}
\value{
A \code{data.frame} describing a CAVE-table related to the BANC project.
In the case of \code{banc_cave_tables}, a vector is returned containing the names of
all queryable cave tables.
}
\description{
Read BANC CAVE-tables, good sources of metadata
}
\examples{
\donttest{
all_banc_soma_positions <- banc_nuclei()
points3d(nat::xyzmatrix(all_banc_soma_positions))
}
}
\seealso{
\code{fafbseg::\link{flywire_cave_query}}
}
