% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave-tables.R
\name{banc_codex_annotations}
\alias{banc_codex_annotations}
\title{Read BANC-FlyWireCodex annotation table}
\usage{
banc_codex_annotations(live = TRUE, ...)
}
\arguments{
\item{live}{logical, get the most recent data or pull from the latest materialisation}

\item{...}{method passed to \code{\link{banc_cave_query}}.}

\item{rootids}{#' @param rootids Character vector specifying one or more BANC rootids. As a
convenience this argument is passed to \code{\link{banc_ids}} allowing you
to pass in data.frames, BANC URLs or simple ids.}
}
\value{
A \code{data.frame} describing that should be similar to what you find for BANC
in FlyWireCodex.
}
\description{
Read BANC-FlyWireCodex annotation table
}
\examples{
\dontrun{
banc.meta <- banc_codex_annotations()
}
}
\seealso{
\code{\link{banc_cave_tables}}
}
