% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{banc_decapitate}
\alias{banc_decapitate}
\title{Subset points to be in the brain or in the VNC}
\usage{
banc_decapitate(x, invert = FALSE, reference = "BANC")
}
\arguments{
\item{x}{an object with 3d points to be subsetted, e.g. an xyz matrix, a \code{neuron}, \code{neuronlist} or a \code{mesh3d} object.
Points must be in native BANC space, i.e. plottable inside \code{banc.surf}.}

\item{invert}{if \code{FALSE} returns brain points, if \code{TRUE} returns VNC points.}
}
\value{
Remove points above or below the midsection of the neck connective of BANC.
}
\description{
Subset points to be in the brain or in the VNC
}
\examples{
\donttest{
m = read_banc_meshes("720575941650432785")
m.brain = banc_decapitate(m)
m.vnc = banc_decapitate(m, invert = TRUE)
plot3d(m.brain, col = "red")
plot3d(m.brain, col = "cyan")
plot3d(banc.surf, col = "grey", alpha = 0.1)
}
}
\seealso{
\code{\link{banc.surf}}
}
