% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{banc_influence_path}
\alias{banc_influence_path}
\title{Resolve local path to influence parquet files, downloading from GCS if needed}
\usage{
banc_influence_path(local_path = NULL, force_download = FALSE)
}
\description{
Resolve local path to influence parquet files, downloading from GCS if needed
}
\keyword{internal}
