% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{banc_read_neuron_meshes}
\alias{banc_read_neuron_meshes}
\title{Read one or more BANC neuron and nuclei meshes}
\usage{
banc_read_neuron_meshes(ids, savedir = NULL, format = c("ply", "obj"), ...)
}
\arguments{
\item{ids}{One or more root ids}

\item{savedir}{An optional location to save downloaded meshes. This acts as a
simple but effective cache since flywire neurons change id whenever they
are edited.}

\item{format}{whether to save meshes in Wavefront obj or Stanford poly
format. obj is the default but ply is a simpler and more compact format.}

\item{...}{Additional arguments passed to
\code{fafbseg::\link{read_cloudvolume_meshes}}}

\item{lod}{The level of detail (highest resolution is 0, default of 2 gives a
good overall morphology while 3 is also useful and smaller still).}

\item{method}{How to treat the mesh object returned from neuroglancer, i.e. as
a \code{mesh3d} object or a \code{ply} mesh.}
}
\value{
A \code{\link[nat]{neuronlist}} containing one or more \code{mesh3d}
objects. See \code{nat::\link[nat]{read.neurons}} for details.
}
\description{
Read one or more BANC neuron and nuclei meshes
}
\examples{
\donttest{
neuron.mesh <- banc_read_neuron_meshes("720575941650432785")
plot3d(neuron.mesh, alpha = 0.1)
nucleus.mesh <- banc_read_neuron_meshes("72903876004544795")
plot3d(nucleus.mesh, col = "black")
}
}
\seealso{
\code{fafbseg::\link{read_cloudvolume_meshes}}
}
