% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2.R
\name{banc_reroot}
\alias{banc_reroot}
\alias{banc_reroot.neuron}
\alias{banc_reroot.neuronlist}
\title{Re-root BANC neuron skeleton at soma}
\usage{
banc_reroot(
  x,
  id = NULL,
  banc_nuclei = bancr::banc_nuclei(),
  estimate = TRUE,
  ...
)

\method{banc_reroot}{neuron}(
  x,
  id = NULL,
  banc_nuclei = bancr::banc_nuclei(),
  estimate = TRUE,
  ...
)

\method{banc_reroot}{neuronlist}(
  x,
  id = NULL,
  banc_nuclei = bancr::banc_nuclei(),
  estimate = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{banc.neurite} object representing the neuron skeleton.}

\item{id}{(Optional) The \code{root_id} of the neuron in the \code{banc_nuclei} data
frame. If NULL, it will be taken from the \code{x$root_id} slot.}

\item{banc_nuclei}{A data frame containing information about nuclei
obtained using \code{bancr::banc_nuclei()}. This data frame is assumed to have
columns named \code{root_id} and \code{nucleus_position_nm}, where \code{nucleus_position_nm}
specifies the 3D coordinates of the soma for each \code{root_id}.}

\item{estimate}{if \code{TRUE} and nucleus position is not in \code{banc_nuclei},
then root is estimated as a leaf node furthest outside of the brain neuropil.}

\item{...}{Methods passed to \code{nat::nlapply}.}
}
\value{
The function returns the re-rooted \code{neuron} object.
}
\description{
This function re-roots a neuron skeleton represented as a \code{neuron}
object at the location of the corresponding soma in the \code{banc_nuclei} data
frame. It uses the \code{root_id} in the skeleton object to identify the soma
location.
}
\examples{
\dontrun{
x <- banc_read_l2skel(..., simplify = FALSE)
banc_nuclei <- banc_nuclei()
re-rooted_neuron <- banc_reroot(x, banc_nuclei = banc_nuclei)
}
}
