% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bancr-package.R
\docType{package}
\name{bancr-package}
\alias{bancr}
\alias{bancr-package}
\title{bancr: R Client Access to the Brain And Nerve Cord (BANC) Dataset}
\description{
Access to the BANC 'righty' EM dataset, focussing on support for BANC autosegmentation. Organised as a wrapper around the fafbseg package.
}
\section{Details}{
 There are several online bancr resources in addition to the
built-in R help documentation. These include:
\itemize{

\item \href{https://flyconnectome.github.io/bancr/}{Package README} including  \href{https://flyconnectome.github.io/bancr/#installation}{installation details} and an \href{https://flyconnectome.github.io/bancr/#ascending-neuron-vignette}{analysis full worked example}.

\item \href{https://flyconnectome.github.io/bancr/reference/index.html}{Function reference}

\item \href{https://flyconnectome.github.io/bancr/SUPPORT.html}{Getting help}

}

Please see below for github source code and issues links.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flyconnectome/bancr}
  \item \url{https://flyconnectome.github.io/bancr/}
  \item Report bugs at \url{https://github.com/flyconnectome/bancr/issues}
}

}
\author{
\strong{Maintainer}: Alexander Bates \email{alexander.shakeel.bates@gmail.com} (\href{https://orcid.org/0000-0002-1195-0445}{ORCID})

Authors:
\itemize{
  \item Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})
}

}
\keyword{internal}
