% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banc-coconatfly.R
\name{register_banc_coconat}
\alias{register_banc_coconat}
\title{Use BANC data with coconat for connectivity similarity}
\usage{
register_banc_coconat(showerror = TRUE)
}
\arguments{
\item{showerror}{Logical: error-out silently or not.}
}
\description{
Register the BANC dataset with \href{https://github.com/natverse/coconat}{coconat},
a natverse R package for between and within dataset connectivity comparisons using cosine similarity.
}
\details{
\code{register_banc_coconat()} registers \code{bancr}-backed functionality for use with
}
\examples{
\dontrun{
library(coconatfly)
banc_meta_create_cache(use_seatable=TRUE)
register_banc_coconat()
cf_cosine_plot(cf_ids('/type:LAL0(08|09|10|42)', datasets = c("banc", "hemibrain")))
}
}
\seealso{
\code{\link[=banc_meta_create_cache]{banc_meta_create_cache()}}
}
